Added Email Features
Added some features to the emails that are sent.

Now any emails that are the same as the previous email sent are marked with COPY at the end.

This is so you can use this in your email app to discard into a differant folder then your main. (so you can still check if the app is working).